Change structure & add a logger
