+++ makefile
@@ -15,4 +15,4 @@ deploy: build
 	@aws lambda update-function-code --function-name Sedy --zip-file fileb://build/sedy.zip
 
 test:
-	@./node_modules/.bin/mocha --compilers js:babel-core/register --require babel-polyfill --require co-mocha --recursive ./lib/**/*.spec.js
+	@./node_modules/.bin/mocha --compilers js:babel-core/register --require babel-polyfill --require co-mocha --recursive ./src/**/*.spec.js
