+++ src/lib/logger.js
@@ -1,0 +1,5 @@
+export default config => ({
+    info: console.log,
+    debug: config.logs.debug ? console.log : () => null,
+    error: console.error,
+});
