+++ src/parser.js
